requirements.txt
amqp==2.1.4
amqplib==1.0.2
anyjson==0.3.3
Babel==0.9.6
billiard==3.5.0.2
celery==4.0.2
crontab==0.21.3
defusedxml==0.5.0
Django==1.9.12
django-annoying==0.10.3
django-filter==1.0.1
django-haystack==2.6.0
django-react-templatetags==2.4.0
django-webpack-loader==0.4.1
djangorestframework==3.5.4
gunicorn==19.0.0
jira==1.0.10
kombu==4.0.2
oauthlib==2.0.2
ordereddict==1.1
pbr==2.0.0
pytz==2016.10
redis==2.8.0
requests==2.13.0
requests-oauthlib==0.8.0
requests-toolbelt==0.7.1
six==1.10.0
vine==1.1.3
whitenoise==3.3.0
